fix: deleteCard

+++ controllers/cards.js
@@ -13,25 +13,39 @@ module.exports.getAllCards = (req, res) => {
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(() => {
-      const error = new Error("Нет карточки по заданному id");
-      error.statusCode = ERROR_NOTFOUND;
-      throw error;
-    })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === "CastError") {
-        res.status(ERROR_BADREQUEST).send({
-          message: "Карточка с указанным _id не найдена.",
+  const { cardId } = req.params;
+
+  Card.findById(cardId).then((card) => {
+    if (!card) {
+      res.status(ERROR_BADREQUEST).send({
+        message: "Карточка с указанным _id не найдена.",
+      });
+    } else if (JSON.stringify(req.user._id) === JSON.stringify(card.owner)) {
+      Card.findByIdAndRemove(req.params.cardId)
+        .orFail(() => {
+          const error = new Error("Нет карточки по заданному id");
+          error.statusCode = ERROR_NOTFOUND;
+          throw error;
+        })
+        .then((delCard) => res.send({ data: delCard }))
+        .catch((err) => {
+          if (err.name === "CastError") {
+            res.status(ERROR_BADREQUEST).send({
+              message: "Карточка с указанным _id не найдена.",
+            });
+          } else if (err.statusCode === ERROR_NOTFOUND) {
+            res.status(ERROR_NOTFOUND).send({
+              message: err.message,
+            });
+          }
+          res.status(ERROR_DEFAULT).send({ message: "Произошла ошибка" });
         });
-      } else if (err.statusCode === ERROR_NOTFOUND) {
-        res.status(ERROR_NOTFOUND).send({
-          message: err.message,
-        });
-      }
-      res.status(ERROR_DEFAULT).send({ message: "Произошла ошибка" });
-    });
+    } else {
+      res.status(ERROR_BADREQUEST).send({
+        message: "нельзя удалять чужие карточки",
+      });
+    }
+  });
 };
 
 module.exports.createCard = (req, res) => {
